Make flags in Java a 16-bit integer like in C
* This can save up memory when it is in the same 4 bytes as the `boolean newLineFlag`

+++ templates/template.rb
@@ -108,7 +108,7 @@ end
 # convenient methods for checking if a flag is set.
 class FlagsParam < Param
   def rbs_class = "Integer"
-  def java_type = "int"
+  def java_type = "short"
   def kind = options.fetch(:kind)
 end
